+++ ext/opentelemetry-ext-opencensusexporter/tests/test_otcollector_metrics_exporter.py
@@ -23,8 +23,8 @@ from opentelemetry import metrics
 from opentelemetry.ext.opencensusexporter import metrics_exporter
 from opentelemetry.sdk.metrics import (
     Counter,
-    Measure,
     MeterProvider,
+    ValueRecorder,
     get_labels_as_key,
 )
 from opentelemetry.sdk.metrics.export import (
@@ -76,7 +76,7 @@ class TestCollectorMetricsExporter(unittest.TestCase):
         )
         self.assertIs(result, metrics_pb2.MetricDescriptor.CUMULATIVE_DOUBLE)
         result = metrics_exporter.get_collector_metric_type(
-            Measure("testName", "testDescription", "unit", None, None)
+            ValueRecorder("testName", "testDescription", "unit", None, None)
         )
         self.assertIs(result, metrics_pb2.MetricDescriptor.UNSPECIFIED)
 
@@ -88,8 +88,8 @@ class TestCollectorMetricsExporter(unittest.TestCase):
         float_counter = self._meter.create_metric(
             "testName", "testDescription", "unit", float, Counter
         )
-        measure = self._meter.create_metric(
-            "testName", "testDescription", "unit", float, Measure
+        valuerecorder = self._meter.create_metric(
+            "testName", "testDescription", "unit", float, ValueRecorder
         )
         result = metrics_exporter.get_collector_point(
             MetricRecord(int_counter, self._key_labels, aggregator)
@@ -106,7 +106,7 @@ class TestCollectorMetricsExporter(unittest.TestCase):
         self.assertRaises(
             TypeError,
             metrics_exporter.get_collector_point(
-                MetricRecord(measure, self._key_labels, aggregator)
+                MetricRecord(valuerecorder, self._key_labels, aggregator)
             ),
         )